Move mind map extractor into knowledge compile package
